chore: move getHiddenPeriods and basemap filtering
