+++ src/components/layers/basemaps/BasemapList.js
@@ -1,7 +1,6 @@
 import { useCachedDataQuery } from '@dhis2/analytics'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { layerTypes } from '../../map/MapApi.js'
 import Basemap from './Basemap.js'
 import styles from './styles/BasemapList.module.css'
 
@@ -9,16 +8,14 @@ const BasemapList = ({ selectedID, selectBasemap }) => {
     const { basemaps } = useCachedDataQuery()
     return (
         <div className={styles.basemapList} data-test="basemaplist">
-            {basemaps
-                .filter((basemap) => layerTypes.includes(basemap.config.type))
-                .map((basemap, index) => (
-                    <Basemap
-                        key={`basemap-${index}`}
-                        onClick={selectBasemap}
-                        isSelected={basemap.id === selectedID}
-                        {...basemap}
-                    />
-                ))}
+            {basemaps.map((basemap, index) => (
+                <Basemap
+                    key={`basemap-${index}`}
+                    onClick={selectBasemap}
+                    isSelected={basemap.id === selectedID}
+                    {...basemap}
+                />
+            ))}
         </div>
     )
 }
